separate rd2sql
